Random exploration framework initialization. Definition of states
Discretized time and space for the definition of states, created functions to choose actions randomly.
Random exploration can almost be performed, there is still a problem in the evaluation of the direction.
The whole code should be looked at and modified so that it handles arrays or np arrays instead of tuples.